# docs/openapi.yml
openapi: "3.1.0"
info:
  title: "TicketRushServer API"
  description: "TicketRushServer API"
  version: "1.0.2"
servers:
  - url: "https://TicketRushServer"
paths:
...
  /v1/public/alias/{alias}:
    get:
      summary: "GET v1/public/alias/{alias}"
      operationId: "getOrganizationInfoByAlias"
      parameters:
        - name: "alias"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PublicOrganizationInfo"
  /v1/public/event/{id}:
    get:
      summary: "GET v1/public/event/{id}"
      operationId: "getEventInfo"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PublicEventInfo"
  /v1/public/org/{id}:
    get:
      summary: "GET v1/public/org/{id}"
      operationId: "getOrganizationInfo"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PublicOrganizationInfo"
...
  /v1/organization:
    get:
      summary: "GET v1/organization"
      operationId: "getInfo"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FullOrganizationInfo"
    patch:
      summary: "PATCH v1/organization"
      operationId: "updateInfo"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationInfoPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/avatar:
    put:
      summary: "PUT v1/organization/avatar"
      operationId: "updateAvatar"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: "object"
              required:
                - "file"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
...
  /v1/purchase/event/{eventId}:
    get:
      summary: "GET v1/purchase/event/{eventId}"
      operationId: "getPurchaseEvent"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PurchaseEventView"
  /v1/purchase/event/{eventId}/seats:
    get:
      summary: "GET v1/purchase/event/{eventId}/seats"
      operationId: "getSeatStatuses"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SeatStatusCollectionView"
  /v1/purchase/hold:
    post:
      summary: "POST v1/purchase/hold"
      operationId: "createHold"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHoldPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HoldView"
  /v1/purchase/hold/{holdId}:
    get:
      summary: "GET v1/purchase/hold/{holdId}"
      operationId: "getHold"
      parameters:
        - name: "holdId"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HoldView"
    delete:
      summary: "DELETE v1/purchase/hold/{holdId}"
      operationId: "releaseHold"
      parameters:
        - name: "holdId"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/purchase/hold/{holdId}/seat:
    post:
      summary: "POST v1/purchase/hold/{holdId}/seat"
      operationId: "addSeatToHold"
      parameters:
        - name: "holdId"
          in: "path"
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoldSeatPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HoldView"
  /v1/purchase/pay/{holdId}:
    post:
      summary: "POST v1/purchase/pay/{holdId}"
      operationId: "completeMockPayment"
      parameters:
        - name: "holdId"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompletedPurchaseView"
  /v1/user:
    get:
      summary: "GET v1/user"
      operationId: "getInfo"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FullUserInfo"
    patch:
      summary: "PATCH v1/user"
      operationId: "updateInfo"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInfoPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/user/avatar:
    put:
      summary: "PUT v1/user/avatar"
      operationId: "updateAvatar"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: "object"
              required:
                - "file"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/user/follow/{id}:
    put:
      summary: "PUT v1/user/follow/{id}"
      operationId: "followOrganization"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
    delete:
      summary: "DELETE v1/user/follow/{id}"
      operationId: "unfollowOrganization"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
...
components:
  schemas:
...
    OperationResult:
      type: "object"
      properties:
        success:
          type: "boolean"
          nullable: true
        message:
          type: "string"
          nullable: true
        resourceId:
          type: "integer"
          format: "int64"
          nullable: true
...
    Instant:
      type: "object"
      properties: { }
...
    PublicOrganizationInfo:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        name:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
        aliasName:
          type: "string"
          nullable: true
        avatarUrl:
          type: "string"
          nullable: true
        bannerUrl:
          type: "string"
          nullable: true
        websiteUrl:
          type: "string"
          nullable: true
        verified:
          type: "boolean"
          nullable: true
        followerCount:
          type: "integer"
          format: "int64"
          nullable: true
    PublicEventInfo:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        name:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
        isOnlineEvent:
          type: "boolean"
          nullable: true
        venue:
          type: "string"
          nullable: true
        address:
          type: "string"
          nullable: true
        dateTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
        bannerUrl:
          type: "string"
          nullable: true
        organizationId:
          type: "integer"
          format: "int64"
          nullable: true
        organizationName:
          type: "string"
          nullable: true
        organizationAlias:
          type: "string"
          nullable: true
...
    FullOrganizationInfo:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        name:
          type: "string"
          nullable: true
        email:
          type: "string"
          nullable: true
        createdAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
        updatedAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
        verified:
          type: "boolean"
          nullable: true
        description:
          type: "string"
          nullable: true
        aliasName:
          type: "string"
          nullable: true
        avatarUrl:
          type: "string"
          nullable: true
        bannerUrl:
          type: "string"
          nullable: true
        websiteUrl:
          type: "string"
          nullable: true
    UpdateOrganizationInfoPayload:
      type: "object"
      properties:
        description:
          type: "string"
          nullable: true
        aliasName:
          type: "string"
          nullable: true
        websiteUrl:
          type: "string"
          nullable: true
...
    SeatView:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        index:
          type: "integer"
          format: "int32"
          nullable: true
        number:
          type: "integer"
          format: "int32"
          nullable: true
        availability:
          type: "string"
          nullable: true
          enum:
            - "AVAILABLE"
            - "SOLD"
            - "HELD_BY_ME"
            - "HELD"
    SeatRowView:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        index:
          type: "integer"
          format: "int32"
          nullable: true
        label:
          type: "string"
          nullable: true
        seats:
          type: "array"
          nullable: true
          items:
            $ref: "#/components/schemas/SeatView"
...
    SalesRoundView:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        name:
          type: "string"
          nullable: true
        startTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
        endTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
        maxTicketsPerPurchase:
          type: "integer"
          format: "int32"
          nullable: true
    TicketClassView:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        name:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
        price:
          type: "integer"
          format: "int64"
          nullable: true
        salesRoundId:
          type: "integer"
          format: "int64"
          nullable: true
        seatZoneId:
          type: "integer"
          format: "int64"
          nullable: true
    SeatZoneView:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        name:
          type: "string"
          nullable: true
        positionX:
          type: "integer"
          format: "int32"
          nullable: true
        positionY:
          type: "integer"
          format: "int32"
          nullable: true
        rows:
          type: "array"
          nullable: true
          items:
            $ref: "#/components/schemas/SeatRowView"
    ActiveHoldView:
      type: "object"
      properties:
        holdId:
          type: "string"
          nullable: true
        expiresAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
        totalAmount:
          type: "integer"
          format: "int64"
          nullable: true
    PurchaseEventView:
      type: "object"
      properties:
        eventId:
          type: "integer"
          format: "int64"
          nullable: true
        eventName:
          type: "string"
          nullable: true
        eventDateTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
        salesRounds:
          type: "array"
          nullable: true
          items:
            $ref: "#/components/schemas/SalesRoundView"
        ticketClasses:
          type: "array"
          nullable: true
          items:
            $ref: "#/components/schemas/TicketClassView"
        seatZones:
          type: "array"
          nullable: true
          items:
            $ref: "#/components/schemas/SeatZoneView"
        myActiveHold:
          nullable: true
          $ref: "#/components/schemas/ActiveHoldView"
    SeatStatusCollectionView:
      type: "object"
      properties:
        eventId:
          type: "integer"
          format: "int64"
          nullable: true
        seatZones:
          type: "array"
          nullable: true
          items:
            $ref: "#/components/schemas/SeatZoneView"
        myActiveHold:
          nullable: true
          $ref: "#/components/schemas/ActiveHoldView"
    CreateHoldPayload:
      type: "object"
      properties:
        eventId:
          type: "integer"
          format: "int64"
          nullable: true
    HeldItemView:
      type: "object"
      properties:
        seatId:
          type: "integer"
          format: "int64"
          nullable: true
        ticketClassId:
          type: "integer"
          format: "int64"
          nullable: true
        price:
          type: "integer"
          format: "int64"
          nullable: true
    HoldView:
      type: "object"
      properties:
        holdId:
          type: "string"
          nullable: true
        expiresAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
        totalAmount:
          type: "integer"
          format: "int64"
          nullable: true
        items:
          type: "array"
          nullable: true
          items:
            $ref: "#/components/schemas/HeldItemView"
    HoldSeatPayload:
      type: "object"
      properties:
        seatId:
          type: "integer"
          format: "int64"
          nullable: true
        ticketClassId:
          type: "integer"
          format: "int64"
          nullable: true
    CompletedPurchaseView:
      type: "object"
      properties:
        purchaseId:
          type: "integer"
          format: "int64"
          nullable: true
        amount:
          type: "integer"
          format: "int64"
          nullable: true
        ticketIds:
          type: "array"
          nullable: true
          items:
            type: "integer"
            format: "int64"
    FullUserInfo:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        name:
          type: "string"
          nullable: true
        email:
          type: "string"
          nullable: true
        avatarUrl:
          type: "string"
          nullable: true
        birthDate:
          type: "string"
          format: "date-time"
          nullable: true
        gender:
          type: "string"
          nullable: true
        phoneNumber:
          type: "string"
          nullable: true
        addressLine:
          type: "string"
          nullable: true
        createdAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
        updatedAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
    UpdateUserInfoPayload:
      type: "object"
      properties:
        birthDate:
          type: "string"
          format: "date-time"
          nullable: true
        gender:
          type: "string"
          nullable: true
          enum:
            - "MALE"
            - "FEMALE"
            - "OTHER"
        phoneNumber:
          type: "string"
          nullable: true
        addressLine:
          type: "string"
          nullable: true
...